Extract automation DialogCreator to standalone file
