+++ Gui/FileDialogs/FileDialog.cs
@@ -23,86 +23,6 @@ namespace MatterHackers.Agg.UI
 		public abstract string ResolveFilePath(string path);
 	}
 
-	internal class AggFileDialogCreator : FileDialogCreator
-	{
-		internal AggFileDialogCreator()
-		{
-		}
-
-		public override bool OpenFileDialog(OpenFileDialogParams openParams, OpenFileDialogDelegate callback)
-		{
-			ShowFileDialog((fileText) =>
-			{
-				if (fileText.Length > 2)
-				{
-					string[] files = fileText.Split(';', ' ').Select(f => f.Trim('\"')).ToArray();
-					openParams.FileName = files[0];
-					openParams.FileNames = files;
-				}
-				UiThread.RunOnIdle(() => callback?.Invoke(openParams));
-			});
-
-			return true;
-		}
-
-		public override bool SaveFileDialog(SaveFileDialogParams saveParams, SaveFileDialogDelegate callback)
-		{
-			ShowFileDialog((fileText) =>
-			{
-				if (fileText.Length > 2)
-				{
-					string[] files = fileText.Split(';', ' ').Select(f => f.Trim('\"')).ToArray();
-					saveParams.FileName = files[0];
-					saveParams.FileNames = files;
-				}
-				UiThread.RunOnIdle(() => callback?.Invoke(saveParams));
-			});
-
-			return true;
-		}
-
-		private static void ShowFileDialog(Action<string> dialogClosedHandler)
-		{
-			var systemWindow = new SystemWindow(600, 200)
-			{
-				Title = "TestAutomation File Input",
-				BackgroundColor = RGBA_Bytes.DarkGray
-			};
-
-			var warningLabel = new TextWidget("This dialog should not appear outside of automation tests.\nNotify technical support if visible", pointSize: 15, textColor: RGBA_Bytes.Pink)
-			{
-				Margin = new BorderDouble(20),
-				VAnchor = VAnchor.ParentTop,
-				HAnchor = HAnchor.ParentLeftRight
-			};
-			systemWindow.AddChild(warningLabel);
-
-			var fileNameInput = new TextEditWidget(pixelWidth: 400)
-			{
-				VAnchor = VAnchor.ParentCenter,
-				HAnchor = HAnchor.ParentLeftRight,
-				Margin = new BorderDouble(30, 15)
-			};
-			fileNameInput.EnterPressed += (s, e) => systemWindow.CloseOnIdle();
-			systemWindow.AddChild(fileNameInput);
-
-			systemWindow.Load += (s, e) => fileNameInput.Focus();
-			systemWindow.Closed += (s, e) =>
-			{
-				dialogClosedHandler(fileNameInput.Text);
-			};
-
-			systemWindow.ShowAsSystemWindow();
-		}
-
-		public override bool SelectFolderDialog(SelectFolderDialogParams folderParams, SelectFolderDialogDelegate callback)
-		{
-			throw new NotImplementedException();
-		}
-
-		public override string ResolveFilePath(string path) => path;
-	}
-
 	public static class FileDialog
 	{
 		private static string lastDirectoryUsed = "";
@@ -120,7 +40,7 @@ namespace MatterHackers.Agg.UI
 
 					if (fileDialogCreatorPlugins.Plugins.Count == 0)
 					{
-						fileDialogCreatorPlugin = new AggFileDialogCreator();
+						fileDialogCreatorPlugin = new AutomationFileDialogCreator();
 					}
 					else
 					{
